refactor: reorganize the structure of the codebase
