+++ src/data_display/grid_plotting_legacy.py
@@ -3,7 +3,7 @@ from typing import Optional
 import pandapower as pp
 import polars as pl
 from polars import col as c
-from helper_functions import pl_to_dict
+from helpers import pl_to_dict
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 import igraph as ig
@@ -14,7 +14,7 @@ import plotly.colors
 from PIL import ImageColor
 from _plotly_utils.basevalidators import ColorscaleValidator
 
-from helper_functions import generate_tree_graph_from_edge_data
+from helpers import generate_tree_graph_from_edge_data
 from data_model import NodeEdgeModel4Reconfiguration
 from pipelines.reconfiguration import DigAPlan, DigAPlanADMM
