+++ src/helpers/__init__.py
@@ -1,4 +1,4 @@
-from helper_functions.general_functions import (
+from helpers.general import (
     pl_to_dict,
     build_non_existing_dirs,
     generate_log,
@@ -7,19 +7,19 @@ from helper_functions.general_functions import (
     dict_to_duckdb,
     duckdb_to_dict,
 )
-from helper_functions.polars_functions import (
+from helpers.polars import (
     get_transfo_impedance,
     get_transfo_imaginary_component,
     cast_boolean,
     modify_string_col,
 )
-from helper_functions.networkx_functions import (
+from helpers.networkx import (
     generate_tree_graph_from_edge_data,
     generate_nx_edge,
     generate_bfs_tree_with_edge_data,
     get_all_edge_data,
 )
-from helper_functions.shapely_functions import load_shape_from_geo_json
+from helpers.shapely import load_shape_from_geo_json
 
 __all__ = [
     "pl_to_dict",
