+++ src/pipeline_reconfiguration/model_managers/combined.py
@@ -1,6 +1,6 @@
 import random
 import pyomo.environ as pyo
-from helper_functions import generate_log
+from helpers import generate_log
 from pipelines.reconfiguration.data_manager import PipelineDataManager
 from data_model.reconfiguration_configs import CombinedConfig
 from model_reconfiguration import generate_combined_model, generate_combined_lin_model
